% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_parent_interval.R
\name{add_parent_interval}
\alias{add_parent_interval}
\alias{add_parent_interval.default}
\alias{add_parent_interval.data.frame}
\alias{add_parent_interval.data.table}
\alias{add_parent_interval.tbl_df}
\title{Calculate parent intervals}
\usage{
add_parent_interval(...)

\method{add_parent_interval}{default}(
  id,
  start,
  end,
  ...,
  name_id = "id",
  name_parent_start = "parent_start",
  name_parent_end = "parent_end",
  name_interval_number = "per_id_interval_number"
)

\method{add_parent_interval}{data.frame}(
  x,
  id,
  start,
  end,
  ...,
  name_parent_start = "parent_start",
  name_parent_end = "parent_end",
  name_interval_number = "per_id_interval_number"
)

\method{add_parent_interval}{data.table}(
  x,
  id,
  start,
  end,
  ...,
  name_parent_start = "parent_start",
  name_parent_end = "parent_end",
  name_interval_number = "per_id_interval_number"
)

\method{add_parent_interval}{tbl_df}(
  x,
  id,
  start,
  end,
  ...,
  name_parent_start = "parent_start",
  name_parent_end = "parent_end",
  name_interval_number = "per_id_interval_number"
)
}
\arguments{
\item{...}{Further arguments passed to or from other methods.}

\item{id}{For the default method, vector representing the id associated with the
episode.

For the data frame methods. a variable in \code{x} representing the id associated
with an episode.}

\item{start}{\verb{[character]}.

For the default method, vector representing episode start date/time.

For the data frame method, a variable in \code{x} representing the episode start.

Variable must be a \verb{<Date>} or \verb{<POSIXct>} object.}

\item{end}{\verb{[character]}

For the default method, vector representing episode end date/time.

For the data frame method, a variable in \code{x} representing the episode end.

Variable must be a \verb{<Date>} or \verb{<POSIXct>} object.

Must refer to a variable that is the same class as \code{start}.}

\item{name_id}{The column name to use for the patient id in the output.}

\item{name_parent_start}{The column name to use for the start of the parent interval in the output.}

\item{name_parent_end}{The column name to use for the end of the parent interval in the output.}

\item{name_interval_number}{The column name to use for the interval number within the matched parent
interval in the output.}

\item{x}{Data frame like object.}
}
\value{
The input data with additional columns for the corresponding parent interval
(split across the \code{id} values).

The returned object will be of the same class as the input \code{x} (i.e. a
data.frame, data.table or tibble).
}
\description{
\code{add_parent_interval()} calculates the minimum spanning interval that
contains overlapping episodes and adds this to the input.
}
\examples{
\dontshow{.old <- data.table::setDTthreads(2)}
dat <- data.frame(
    id = c(1, 1, 2, 2, 2, 1),
    start = as.Date(c(
        "2020-01-01", "2020-01-03", "2020-04-01",
        "2020-04-15", "2020-04-17", "2020-05-01"
    )),
    end = as.Date(c(
        "2020-01-10", "2020-01-10", "2020-04-30",
        "2020-04-16", "2020-04-19", "2020-10-01"
    ))
)

with(dat, add_parent_interval(id, start, end))
add_parent_interval(dat, id = "id", start = "start", end = "end")
\dontshow{data.table::setDTthreads(.old)}

}
