% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_episodes.R
\name{merge_episodes}
\alias{merge_episodes}
\alias{merge_episodes.default}
\alias{merge_episodes.data.frame}
\alias{merge_episodes.data.table}
\alias{merge_episodes.tbl_df}
\title{Merge overlapping episodes}
\usage{
merge_episodes(...)

\method{merge_episodes}{default}(
  id,
  start,
  end,
  ...,
  name_id = "id",
  name_episode_start = "episode_start",
  name_episode_end = "episode_end",
  name_episode_number = "episode_number"
)

\method{merge_episodes}{data.frame}(
  x,
  id,
  start,
  end,
  ...,
  name_episode_start = "episode_start",
  name_episode_end = "episode_end",
  name_episode_number = "episode_number"
)

\method{merge_episodes}{data.table}(
  x,
  id,
  start,
  end,
  ...,
  name_episode_start = "episode_start",
  name_episode_end = "episode_end",
  name_episode_number = "episode_number"
)

\method{merge_episodes}{tbl_df}(x, id = "id", start = "start", end = "end", ...)
}
\arguments{
\item{...}{Further arguments passed to or from other methods.}

\item{id}{For the default method, vector representing the id associated with the
episode.

For the data frame method, a variable in \code{x} representing the id associated
with the episode.}

\item{start}{For the default method, vector representing episode start date/time.

For the data frame method, a variable in \code{x} representing the episode start.

Variable must be a \verb{<Date>} or \verb{<POSIXct>} object.}

\item{end}{For the default method, vector representing episode end date/time.

For the data frame method, a variable in \code{x} representing the episode end.

Variable must be a \verb{<Date>} or \verb{<POSIXct>} object.

Must refer to a variable that is the same class as \code{start}.}

\item{name_id}{The column name to use for the patient id in the output.}

\item{name_episode_start}{The column name to use for the start of the episode in the output.}

\item{name_episode_end}{The column name to use for the end of the episode in the output.}

\item{name_episode_number}{The column name to use for the episode number of an individual patient in the
output.}

\item{x}{Data frame like object.}
}
\value{
The resulting combined episode intervals split by id and ordered by interval
number.

The returned object will be of the same class as the input \code{x} (i.e. a
data.frame, data.table or tibble).
}
\description{
\code{merge_episodes()} combines overlapping episodes in to a minimal spanning
interval split by in individual identifier.
}
\examples{
\dontshow{.old <- data.table::setDTthreads(2)}
dat <- data.frame(
    id = c(1, 1, 2, 2, 2, 1),
    start = as.Date(c(
        "2020-01-01", "2020-01-03", "2020-04-01",
        "2020-04-15", "2020-04-17", "2020-05-01"
    )),
    end = as.Date(c(
        "2020-01-10", "2020-01-10", "2020-04-30",
        "2020-04-16", "2020-04-19", "2020-10-01"
    ))
)

with(dat, merge_episodes(id, start, end))
merge_episodes(dat, id = "id", start = "start", end = "end")
\dontshow{data.table::setDTthreads(.old)}

}
