% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHSRepisodes-package.R
\docType{package}
\name{NHSRepisodes-package}
\alias{NHSRepisodes}
\alias{NHSRepisodes-package}
\title{NHSRepisodes: Merge Overlapping Episode Intervals}
\description{
Merge overlapping episode intervals.
}
\author{
\strong{Maintainer}: NHS R \email{nhs.rcommunity@nhs.net}

Authors:
\itemize{
  \item Tim Taylor \email{tim.taylor@hiddenelephants.co.uk} (\href{https://orcid.org/0000-0002-8587-7113}{ORCID})
}

}
\keyword{internal}
