% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_parent_interval.R
\name{add_parent_interval}
\alias{add_parent_interval}
\alias{add_parent_interval.default}
\alias{add_parent_interval.data.table}
\alias{add_parent_interval.tbl_df}
\alias{add_parent_interval.data.frame}
\title{Calculate parent intervals}
\usage{
add_parent_interval(x, id = "id", start = "start", end = "end", ...)

\method{add_parent_interval}{default}(x, ...)

\method{add_parent_interval}{data.table}(x, id = "id", start = "start", end = "end", ...)

\method{add_parent_interval}{tbl_df}(x, id = "id", start = "start", end = "end", ...)

\method{add_parent_interval}{data.frame}(x, id = "id", start = "start", end = "end", ...)
}
\arguments{
\item{x}{An \R object.}

\item{id}{\verb{[character]}
Variable in \code{x} representing the id associated with an episode.}

\item{start}{\verb{[character]}
Variable in \code{x} representing the start of the episode. Must refer to a
variable that is either class \verb{<Date>} or \verb{<POSIXct>}.}

\item{end}{\verb{[character]}
Variable in \code{x} representing the start of the episode. Must refer to a
variable that is the same class as \code{start}.}

\item{...}{Not currently used.}
}
\value{
The input data with additional columns for the corresponding parent interval.
The returned object will be of the same class as the input \code{x} (i.e. a
data.frame, data.table or tibble).
}
\description{
\code{add_parent_interval()} calculates the minimum spanning interval that
contains overlapping episodes and adds this to the input. Methods are
provided for data.frame like objects.
}
\examples{
dat <- data.frame(
  id = c(1, 1, 2, 2, 2, 1),
  start = as.Date(c(
    "2020-01-01", "2020-01-03", "2020-04-01",
    "2020-04-15", "2020-04-17", "2020-05-01"
  )),
  end = as.Date(c(
    "2020-01-10", "2020-01-10", "2020-04-30",
    "2020-04-16", "2020-04-19", "2020-10-01"
  ))
)

add_parent_interval(dat)

}
