% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHSRepisodes-package.R
\docType{package}
\name{NHSRepisodes-package}
\alias{NHSRepisodes}
\alias{NHSRepisodes-package}
\title{NHSRepisodes: Package relating to Hospital Episode Intervals}
\description{
Hospital episodes can overlap or have gaps which can result in under or over counting. This package contains functions which can be used to rectify this common analytical issue in NHS data.
}
\author{
\strong{Maintainer}: Zoë Turner \email{zoe.turner3@nhs.net} (\href{https://orcid.org/0000-0003-1033-9158}{ORCID})

Authors:
\itemize{
  \item Tim Taylor \email{tim.taylor@hiddenelephants.co.uk} (\href{https://orcid.org/0000-0002-8587-7113}{ORCID})
}

Other contributors:
\itemize{
  \item NHS R [copyright holder]
}

}
\keyword{internal}
